+++ modules/firefox.nix
@@ -9,9 +9,8 @@
     inherit (pkgs.stdenv) hostPlatform;
   in {
     programs.firefox.enable = true;
-    programs.firefox.package = if (hostPlatform.isDarwin) then
-      # Alert me when Firefox through Nix is supported on macOS
-      assert (pkgs.firefox.meta.unsupported && pkgs.firefox-bin.meta.unsupported); pkgs.emptyDirectory
+    programs.firefox.package = if hostPlatform.isDarwin then
+      pkgs.firefox-bin-unwrapped
     else
       pkgs.firefox;
     programs.firefox.extensions = [